block/partitions/amlogic: Add support
This adds support of Amlogic's proprietary partitions mainly seen on eMMC
of devices using their SoCs to the block subsystem.

The support is based on reverse-engineering so it could have major diffe-
rences on technical details, the most notable one is that Amlogic added
their own support to this format to the MMC driver instead of the block
subsystem.

There are probably multiple iterations of such implementaions by Amlogic
but this support is based on the one with magic MPT and version 01.00.00

It is added as the last parsed partition type after all other existing ones
like msdos, gpt, etc. So any existing msdos or gpt table on a block device
will prevent the Amlogic one from being parsed.

The support could be enabled with CONFIG_AMLOGIC_PARTITION=y

Two kernel command line options could be used to determine the behaviours:
 - apt_checksum= : when set to 0, will turn off checksum; when set to other
 values, or left empty, will turn on checksum
 - apt_blkdevs= : a list of block devices seperated by ',', only devices
 listed here will be parsed. A special value all means all block devices
 should be parsed

Signed-off-by: Guoxin Pu <pugokushin@gmail.com>